Look for assets in subdirectories
Fixes #503.

When running `npm start` the README.skills.md file will be updated to
include any newly-added skills. This includes the name and description
from the skill's frontmatter, but also a list of assets in the skill--
that is, files other than SKILL.md.

However, the code in yaml-parser.mjs to enumerate the asset files only
looks in the immediate skill directory, not any subdirectories. Those
files will be missing from README.skills.md.

Here we update yaml-parser.mjs to recurse through the subdirectories as
well.